requirements.txt
pyosc
requests
